Bleed

+++ OpenRA.Mods.RA/Air/ReturnToBase.cs
@@ -56,17 +56,17 @@ namespace OpenRA.Mods.RA.Air
 			var altitude = aircraft.Altitude;
 			if (altitude == 0) altitude = self.Info.Traits.Get<PlaneInfo>().CruiseAltitude;
 
-			var approachStart = landPos.ToFloat2() - new float2(altitude * speed, 0);
+			var approachStart = landPos.ToInt2() - new float2(altitude * speed, 0);
 			var turnRadius = (128f / self.Info.Traits.Get<AircraftInfo>().ROT) * speed / (float)Math.PI;
 
 			/* work out the center points */
 			var fwd = -float2.FromAngle(aircraft.Facing / 128f * (float)Math.PI);
 			var side = new float2(-fwd.Y, fwd.X);		/* rotate */
 			var sideTowardBase = new[] { side, -side }
-				.OrderBy(a => float2.Dot(a, self.CenterLocation.ToFloat2() - approachStart))
+				.OrderBy(a => float2.Dot(a, self.CenterLocation.ToInt2() - approachStart))
 				.First();
 
-			var c1 = self.CenterLocation.ToFloat2() + turnRadius * sideTowardBase;
+			var c1 = self.CenterLocation.ToInt2() + turnRadius * sideTowardBase;
 			var c2 = approachStart + new float2(0, turnRadius * Math.Sign(self.CenterLocation.Y - approachStart.Y));	// above or below start point
 
 			/* work out tangent points */
